added sensor count by location with test

+++ measurement-storage-service/src/main/java/com/bpm/measurementstorageservice/repository/SensorStorageRepository.java
@@ -9,5 +9,5 @@ public interface SensorStorageRepository extends JpaRepository<Sensor, Long> {
     Optional<Sensor> findBySensorId(String sensorId);
     boolean existsBySensorId(String sensorId);
     Optional<Sensor> findByLocation(String location);
-    long countByLocation(String location);
+    Integer countByLocation(String location);
 }
